requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
pillow
numpy==1.24.1
typing
scikit-image
huggingface_hub
transformers>=4.39.1
gradio